@@ -1,6 +1,4 @@
-// File: postcss.config.js
-
-module.exports = {
+export default {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
