@@ -1,5 +1,3 @@
-// File: morpheus-ide-main/src/components/Tabs.tsx
-
 import React from 'react';
 import { X } from 'lucide-react';
 import { FileType } from '../types';
@@ -11,21 +9,14 @@ interface TabsProps {
   onTabClose: (fileId: string) => void;
 }
 
-export default function Tabs({
-  openFiles,
-  activeFileId,
-  onTabSelect,
-  onTabClose,
-}: TabsProps) {
+export default function Tabs({ openFiles, activeFileId, onTabSelect, onTabClose }: TabsProps) {
   return (
     <div className="flex bg-editor-bg border-b border-editor-border overflow-x-auto">
       {openFiles.map((file) => (
         <div
           key={file.id}
           className={`flex items-center px-4 py-2 border-r border-editor-border min-w-[120px] max-w-[200px] ${
-            activeFileId === file.id
-              ? 'bg-editor-active text-editor-text'
-              : 'text-editor-icon hover:bg-editor-active/50'
+            activeFileId === file.id ? 'bg-editor-active text-editor-text' : 'text-editor-icon hover:bg-editor-active/50'
           }`}
         >
           <button
