@@ -1,8 +1,6 @@
-// File: morpheus-ide-main/src/main.tsx
-
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App';
+import App from './App.tsx';
 import './index.css';
 
 createRoot(document.getElementById('root')!).render(
