@@ -1,12 +1,8 @@
-// File: src/types.ts
-
 export interface FileType {
   id: string;
   name: string;
   content: string;
   language: string;
-  path?: string;
-  handle?: FileSystemFileHandle; // Updated to proper type
 }
 
 export interface FolderType {
@@ -14,14 +10,12 @@ export interface FolderType {
   name: string;
   items: (FileType | FolderType)[];
   isOpen?: boolean;
-  handle?: FileSystemDirectoryHandle; // Updated to proper type
 }
 
-export interface TreeNode {
+export interface Framework {
+  id: string;
   name: string;
-  type: 'file' | 'folder';
-  path?: string;
-  children?: TreeNode[];
+  description: string;
+  icon: JSX.Element;
+  category: 'frontend' | 'backend' | 'fullstack' | 'mobile' | 'ai';
 }
-
-export type Framework = 'React' | 'Flask'; // Add more frameworks as needed
